+++ src/dtagent/plugins/dynamic_tables.sql/admin/032_p_grant_monitor_dynamic_tables.sql
@@ -26,6 +26,7 @@
 --
 -- !! Must be invoked after creation of CONFIG.CONFIGURATIONS table (031_configuration_table)
 --
+--%OPTION:dtagent_admin:
 use role DTAGENT_OWNER; use database DTAGENT_DB; use warehouse DTAGENT_WH;
 
 create or replace procedure DTAGENT_DB.APP.P_GRANT_MONITOR_DYNAMIC_TABLES()
@@ -80,3 +81,4 @@ $$
 ;
 
 grant usage on procedure DTAGENT_DB.APP.P_GRANT_MONITOR_DYNAMIC_TABLES() to role DTAGENT_ADMIN;
+--%:OPTION:dtagent_admin